+++ .gitignore
@@ -6,3 +6,4 @@ node_modules/
 /build/
 /dist/
 /coverage/
+js_out/
