+++ src/rules/no-array-mutation.ts
