+++ src/rules/no-class.ts
