+++ src/rules/no-delete.ts
