+++ src/rules/no-if-statement.ts
