+++ src/rules/no-loop-statement.ts
