+++ src/rules/no-method-signature.ts
