+++ src/rules/no-mixed-interface.ts
