+++ src/rules/no-object-mutation.ts
