+++ src/rules/no-reject.ts
