+++ src/rules/no-throw.ts
