+++ src/rules/no-try.ts
