+++ src/rules/readonly-array.ts
