+++ tests/rules/no-delete.test.ts
@@ -5,7 +5,7 @@
 import dedent from "dedent";
 import { Rule, RuleTester } from "eslint";
 
-import { name, rule } from "../../src/rules/noDelete";
+import { name, rule } from "../../src/rules/no-delete";
 
 import { es3, typescript } from "../configs";
 import {
