+++ tests/rules/no-let.test.ts
@@ -5,7 +5,7 @@
 import dedent from "dedent";
 import { Rule, RuleTester } from "eslint";
 
-import { name, rule } from "../../src/rules/noLet";
+import { name, rule } from "../../src/rules/no-let";
 
 import { es6, typescript } from "../configs";
 import {
