+++ tests/rules/no-mixed-interface.test.ts
@@ -5,7 +5,7 @@
 import dedent from "dedent";
 import { Rule, RuleTester } from "eslint";
 
-import { name, rule } from "../../src/rules/noMixedInterface";
+import { name, rule } from "../../src/rules/no-mixed-interface";
 
 import { typescript } from "../configs";
 import {
