+++ tests/rules/readonly-keyword.test.ts
@@ -5,7 +5,7 @@
 import dedent from "dedent";
 import { Rule, RuleTester } from "eslint";
 
-import { name, rule } from "../../src/rules/readonlyKeyword";
+import { name, rule } from "../../src/rules/readonly-keyword";
 
 import { typescript } from "../configs";
 import {
